[upload-ci-configs] Moved helper tasks to gitlab api
